Created components for DogCards and the home page if a user is logged out.
