+++ client/src/App.js
@@ -12,6 +12,7 @@ import OwnerSettings from "./pages/OwnerSettings";
 // import NotFound from './pages/NotFound';
 import Navbar from './components/Navbar';
 import Footer from './components/Footer';
+// import DogCards from "./components/DogCards";
 
 export default function App() {
   return (
@@ -59,5 +60,3 @@ export default function App() {
     </>
   );
 }
-
-// export default App;
