+++ client/src/components/MainLoggedOut/index.js
@@ -1,0 +1,2 @@
+import React from 'react';
+
